Refactor: session.condition 패키지 분리
