+++ src/main/java/nextstep/courses/domain/session/condition/SessionEnrollmentConditions.java
@@ -1,4 +1,4 @@
-package nextstep.courses.domain.session;
+package nextstep.courses.domain.session.condition;
 
 import nextstep.courses.exception.SessionException;
