+++ src/main/java/nextstep/courses/domain/session/condition/SessionNoneCondition.java
@@ -1,4 +1,4 @@
-package nextstep.courses.domain.session;
+package nextstep.courses.domain.session.condition;
 
 public class SessionNoneCondition implements SessionEnrollmentCondition {
     @Override
